+++ vfssimple/doc.go
@@ -8,31 +8,32 @@ Usage
 
 Just import vfssimple.
 
-  package main
+	package main
 
-  import(
-	"github.com/c2fo/vfs/v6/vfssimple"
-  )
+	import (
+		"fmt"
 
-  ...
+		"github.com/c2fo/vfs/v6/vfssimple"
+	)
 
-  func DoSomething() error {
-    myLocalDir, err := vfssimple.NewLocation("file:///tmp/")
-    if err != nil {
-        return err
-    }
+	func main() {
+		myLocalDir, err := vfssimple.NewLocation("file:///tmp/")
+		if err != nil {
+			panic(err)
+		}
 
-    myS3File, err := vfssimple.NewFile("s3://mybucket/some/path/to/key.txt")
-    if err != nil {
-        return err
-    }
+		myS3File, err := vfssimple.NewFile("s3://mybucket/some/path/to/key.txt")
+		if err != nil {
+			panic(err)
+		}
 
-    localFile, err := myS3File.MoveToLocation(myLocalDir)
-    if err != nil {
-        return err
-    }
+		localFile, err := myS3File.MoveToLocation(myLocalDir)
+		if err != nil {
+			panic(err)
+		}
 
-  }
+		fmt.Printf("moved %s to %s\n", myS3File, localFile)
+	}
 
 Authentication and Options
 
@@ -45,37 +46,70 @@ type/schema with separate credentials, you can register and map file system opti
 The vfssimple library will automatically try to resolve the provided URI in NewFile() or NewLocation() to the registered
 file system.
 
-  package main
+	package main
 
-  import(
-	"github.com/c2fo/vfs/v6/vfssimple"
-	"github.com/c2fo/vfs/v6/backend"
-	"github.com/c2fo/vfs/v6/backend/s3"
-  )
+	import(
+		"fmt"
 
-  ...
+		"github.com/c2fo/vfs/v6/backend"
+		"github.com/c2fo/vfs/v6/backend/s3"
+		"github.com/c2fo/vfs/v6/vfssimple"
+	)
 
-  func DoSomething() error {
-	bucketAuth := s3.NewFileSystem().WithOptions(s3.Options{
-		AccessKeyID:     "key1",
-		SecretAccessKey: "secret1,
-		Region:          "us-west-2",
-	})
+	func main() {
+		bucketAuth := s3.NewFileSystem().WithOptions(s3.Options{
+			AccessKeyID:     "key1",
+			SecretAccessKey: "secret1",
+			Region:          "us-west-2",
+		})
 
-	fileAuth := s3.NewFileSystem().WithOptions(s3.Options{
-		AccessKeyID:     "key2",
-		SecretAccessKey: "secret2,
-		Region:          "us-west-2",
-	})
+		fileAuth := s3.NewFileSystem().WithOptions(s3.Options{
+			AccessKeyID:     "key2",
+			SecretAccessKey: "secret2",
+			Region:          "us-west-2",
+		})
 
-	backend.Register("s3://bucket1/, bucketAuth)
-	backend.Register("s3://bucket2/file.txt, fileAuth)
+		backend.Register("s3://bucket1/", bucketAuth)
+		backend.Register("s3://bucket2/file.txt", fileAuth)
 
-	secureFile, _ := vfssimple.NewFile("s3://bucket2/file.txt")
-	publicLocation, _ := vfssimple.NewLocation("s3://bucket1/")
+		secureFile, _ := vfssimple.NewFile("s3://bucket2/file.txt")
+		publicLocation, _ := vfssimple.NewLocation("s3://bucket1/")
 
-	secureFile.CopyToLocation(publicLocation)
-  }
+		secureFile.CopyToLocation(publicLocation)
+
+		fmt.Printf("copied %s to %s\n", secureFile, publicLocation)
+	}
+
+Registered Backend Resolution
+
+Every backend type automatically registers itself as an available backend filesystem for vfssimple based on its scheme.  In this way,
+vfssimple is able to determine which backend to use for any related URI.  As mentioned above, you can register your own initialized
+filesystem as well.
+
+vfssimple resolves backends by doing a prefix match of the URI to the registered backend names, choosing the longest(most specific) matching
+backend filesystem.
+
+For instance, given registered backends with the names:
+
+	's3'                         - registered by default
+	's3://somebucket/'           - perhaps this was registered using AWS access key id x
+	's3://somebucket/path/'      - and this was registered using AWS access key id y
+	's3://somebucket/path/a.txt' - and this was registered using AWS access key id z
+	's3://some'                  - another contrived registered fs for bucket
+
+See the expected registered bucket name for each:
+
+	's3://somebucket/path/a.txt' - URI: 's3://somebucket/path/a.txt'         (most specific match)
+	's3://somebucket/path/a.txt' - URI: 's3://somebucket/path/a.txt.tar.gz'  (prefix still matches)
+	's3://somebucket/path/'      - URI: 's3://somebucket/path/otherfile.txt' (file only matches path-level registered fs)
+	's3"//somebucket/path/'      - URI: 's3://somebucket/path/'              (exact path-level match)
+	's3://somebucket/'           - URI: 's3://somebucket/test/file.txt'      (bucket-level match only)
+	's3://somebucket/'           - URI: 's3://somebucket/test/'              (still bucket-level match only)
+	's3://somebucket/'           - URI: 's3://somebucket/'                   (exact bucket-level match)
+	's3://some'                  - URI: 's3://some-other-bucket/'            (bucket-level match)
+	's3'                         - URI: 's3://other/'                        (scheme-level match, only)
+	's3'                         - URI: 's3://other/file.txt'                (scheme-level match, only)
+	's3'                         - URI: 's3://other/path/to/nowhere/'        (scheme-level match, only)
 
 */
 package vfssimple