+++ vfssimple/vfssimple.go
@@ -1,16 +1,24 @@
 package vfssimple
 
 import (
+	"errors"
 	"fmt"
 	"net/url"
-	"os"
-	"path/filepath"
 	"strings"
 
 	"github.com/c2fo/vfs/v6"
 	"github.com/c2fo/vfs/v6/backend"
 	_ "github.com/c2fo/vfs/v6/backend/all" // register all backends
 	"github.com/c2fo/vfs/v6/backend/azure"
+	"github.com/c2fo/vfs/v6/backend/mem"
+	"github.com/c2fo/vfs/v6/backend/os"
+)
+
+var (
+	ErrMissingAuthority = errors.New("unable to determine uri authority ([user@]host[:port]) for network-based scheme")
+	ErrMissingScheme    = errors.New("unable to determine uri scheme")
+	ErrRegFsNotFound    = errors.New("no matching registered filesystem found")
+	ErrBlankURI         = errors.New("uri is blank")
 )
 
 // NewLocation is a convenience function that allows for instantiating a location based on a uri string. Any
@@ -19,7 +27,7 @@ import (
 func NewLocation(uri string) (vfs.Location, error) {
 	fs, host, path, err := parseSupportedURI(uri)
 	if err != nil {
-		return nil, err
+		return nil, fmt.Errorf("unable to create vfs.Location for uri %q: %w", uri, err)
 	}
 
 	return fs.NewLocation(host, path)
@@ -31,78 +39,81 @@ func NewLocation(uri string) (vfs.Location, error) {
 func NewFile(uri string) (vfs.File, error) {
 	fs, host, path, err := parseSupportedURI(uri)
 	if err != nil {
-		return nil, err
+		return nil, fmt.Errorf("unable to create vfs.File for uri %q: %w", uri, err)
 	}
 
 	return fs.NewFile(host, path)
 }
 
-func parseSupportedURI(uri string) (vfs.FileSystem, string, string, error) {
-	var err error
+// parseURI attempts to parse a URI and validate that it returns required results
+func parseURI(uri string) (scheme, authority, path string, err error) {
+	// return early if blank uri
+	if uri == "" {
+		err = ErrBlankURI
+		return
+	}
+
+	// parse URI
 	var u *url.URL
 	u, err = url.Parse(uri)
 	if err != nil {
-		return nil, "", "", err
-	}
-	host := u.Host
-	path := u.Path
-	user := u.User
-	if user.String() != "" {
-		host = fmt.Sprintf("%s@%s", user, host)
+		err = fmt.Errorf("unknown url.Parse error: %w", err)
+		return
 	}
 
-	var fs vfs.FileSystem
-	for _, backendScheme := range backend.RegisteredBackends() {
-		// Azure
-		if azure.IsValidURI(u) {
-			host, path, err = azure.ParsePath(u.Path)
-		}
-
-		// Object-level backend
-		if strings.HasPrefix(uri, backendScheme) {
-			fs = backend.Backend(backendScheme)
-			break
-		}
-		// Bucket-level backend
-		volume := fmt.Sprintf("%s://%s/", u.Scheme, u.Host)
-		if volume == backendScheme {
-			fs = backend.Backend(backendScheme)
-			break
-		}
-
-		// Location-level backend
-		if isInPath(u, volume, backendScheme) {
-			fs = backend.Backend(backendScheme)
-			break
-		}
-
-		// Scheme-level backend
-		if u.Scheme == backendScheme {
-			fs = backend.Backend(backendScheme)
-		}
+	// validate schema
+	scheme = u.Scheme
+	if u.Scheme == "" {
+		err = ErrMissingScheme
+		return
 	}
 
-	if fs == nil {
-		err = fmt.Errorf("%s is an unsupported uri scheme", u.Scheme)
+	// validate authority
+	authority = u.Host
+	path = u.Path
+	if azure.IsValidURI(u) {
+		authority, path, err = azure.ParsePath(path)
 	}
 
-	return fs, host, path, err
+	if u.User.String() != "" {
+		authority = fmt.Sprintf("%s@%s", u.User, u.Host)
+	}
+	// network-based schemes require authority, but not file:// or mem://
+	if authority == "" && !(scheme == os.Scheme || scheme == mem.Scheme) {
+		return "", "", "", ErrMissingAuthority
+	}
+
+	return
 }
 
-// isInPath will crawl down the provided url's path to see if it matches the registered root value
-// Example:
-//	 url: s3://bucket/root/path/to/file.txt
-//   root: s3://bucket/root/   <== Registered URI for file system.
-//
-//   This would return true because the target url is within the root.
-func isInPath(u *url.URL, volume, root string) bool {
-	path := u.Path
-	paths := strings.Split(path, string(os.PathSeparator))
-	for i := range paths {
-		path := fmt.Sprintf("%s%s/", volume, filepath.Join(paths[0:i]...))
-		if path == root {
-			return true
+// parseSupportedURI checks if URI matches any backend name as prefix, capturing the longest(most specific) match found.
+// See doc.go Registered Backend Resoltion seciton for examples.
+func parseSupportedURI(uri string) (vfs.FileSystem, string, string, error) {
+	_, authority, path, err := parseURI(uri)
+	if err != nil {
+		return nil, "", "", err
+	}
+
+	var longest string
+	backends := backend.RegisteredBackends()
+	for _, backendName := range backends {
+		if strings.HasPrefix(uri, backendName) {
+			// The first match always becomes the longest
+			if longest == "" {
+				longest = backendName
+				continue
+			}
+
+			// we found a longer (more specific) backend prefix matching URI
+			if len(backendName) > len(longest) {
+				longest = backendName
+			}
 		}
 	}
-	return false
+
+	if longest == "" {
+		err = ErrRegFsNotFound
+	}
+
+	return backend.Backend(longest), authority, path, err
 }